// src/modules/auth/proto-buffers/auth.pb.ts
export interface RegisterResponse {
  status: number;
  error: string[];
}

/** Login */
export interface LoginRequest {
  email: string;
  password: string;
}

export interface LoginResponse {
  status: number;
  error: string[];
  accessToken: string;
  refreshToken: string;
}

/** Refresh token */
export interface RefreshTokenRequest {
  refreshToken: string;
  userId: string;
}

export interface RefreshTokenResponse {
  status: number;
  error: string[];
  accessToken: string;
  refreshToken: string;
}
// ...
export interface ValidateResponse {
  status: number;
  error: string[];
  userId: string;
  role: string;
}